bot/requirements.txt
python_telegram_bot==13.15
redis==4.4.0
docker-compose==1.29.2